moved classe to different folder structure
